requirements.txt
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
scipy==1.7.3
six==1.16.0
